+++ Cpp/source/MulticastReceiver.h
@@ -135,7 +135,7 @@ namespace ops
 				if (!error && nrBytesReceived > 0)
 				{
 					//printf("Data receivedm in multicast receiver\n");
-					handleReadOK(data, nrBytesReceived);
+					handleReadOK(data, (int)nrBytesReceived);
 				}
 				else
 				{
@@ -198,8 +198,8 @@ namespace ops
 		{
 			try
 			{
-				int nReceived = sock->receive_from(boost::asio::buffer(buf, size), lastEndpoint);
-				return nReceived;
+				size_t nReceived = sock->receive_from(boost::asio::buffer(buf, size), lastEndpoint);
+				return (int)nReceived;
 			}
 			catch(...)
 			{
@@ -211,7 +211,7 @@ namespace ops
 
 		int available()
 		{
-			return sock->available();
+			return (int)sock->available();
 		}
 
 		bool sendReply(char* buf, int size)
