+++ Cpp/source/UDPSender.cpp
@@ -127,8 +127,8 @@ namespace ops
 	int UDPSender::receiveReply(char* buf, int size)
 	{
 		if (!socket) return 0;
-		int nReceived = socket->receive(boost::asio::buffer(buf, size));
-		return nReceived;
+		size_t nReceived = socket->receive(boost::asio::buffer(buf, size));
+		return (int)nReceived;
 	}
 	bool UDPSender::waitForReply(int timeout)
 	{
